+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.23.0] - 2025-12-15
+
 ### Changed
 
 - Change `giantswarm.io/helmcharts` entity annotation for components, to include registry hostname and path, like `gsoci.azurecr.io/charts/giantswarm/hello-world`.
@@ -357,7 +359,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial code
 
-[Unreleased]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.22.0...HEAD
+[Unreleased]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.23.0...HEAD
+[0.23.0]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.22.0...v0.23.0
 [0.22.0]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.21.2...v0.22.0
 [0.21.2]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.21.1...v0.21.2
 [0.21.1]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.21.0...v0.21.1